Remove database and feature_id attributes ...
from indicator classes. This information is needed to save and load
results of an indicator to database. This logic should be handled by
oqt.py and geodatabase/client.py, not by the indicator class. In
addition the API does not include those attributes in the response either.

Impulse to change this has been the implementation of async/await for
geodatabase client. The __init__ of the indicator classes relied on
the geodatabase module to get the bpolys if only database and feature_id
has been provided. But the __init__ function of a class can not be
asynchronous.