Save indicator as GeoJSON Feature to DB
Extend database schema with one additional attribute of type JSON.
During saving indicator attribute to database, dump the
indicator.as_feature() output to the database as well. The reason is to
not lose any data when saving and loading and indicator from and to the
database such as the indicator.data attribute (see issue #145).

Workflow to save and load an indicator form the database will stay the same.
This work will be helpful in solving issue #145.