Remove layer "ideal_vgi_lulc" as valid layer for
the Mapping Saturation indicator.

Layer definition "IDEAL-VGI Land Use and Land Cover" has been a valid
layer in combination with the Mapping Saturation indicator.
But often time the area of the land-use gets smaller over time instead
of bigger, which makes it not suitable for the Mapping Saturation.
Also the layer definition is not used in any Report.