add new Indicator about buildings using ML
Add dependency building-completeness-model Python package.
Use this library to preprocess data and make predictions.

Predicts the building area of the AOI using a trained Random Forest
Regressor. The result is the ratio between the prediction of building
area and the building area mapped in OSM.

The input parameters (X or Covariates) to the models are population and
population density (GHSL GHS-POP), settlement typologies (GHSL SMOD),
subnational Humand Development Index (GDL SHDI) and nightlights (EGO
VNL).

The spatial resolution of the model are hex-cells at zoom level 12. The
input AOI is split into hex-cells and the prediction is done for each of
those hex-cells. The model is trained on hex-cells in Africa. Therefor
the Indicator is restricted to input AOI within the bounding box of
Africa.