Update api.md
On the "Request an Indicator for a custom AOI and Layer using Python and `requests` Library" example., the custom Layer call only seems to be supported for the Mapping Saturation indicator. Replaced the GhsPopComparisonBuildings indicator that was originally so this example can work directly. Also added commas that were missing on the layer definition