tests: verify result figures using approval testing
* tests: refactor approvaltests Namer

Move Namer to own module.

Approval files are now stored in `/tests/approvals` instead of
`/tests/integrationtests/approved`.

* fix: add missing whitespace to result description

* tests: use approval testing for result description

Previously tests checked only if result description is not None or
contained a sub-string. Now the whole result description is verified
using approval testing.

* build: add kaleido as dev dependency

to be able to output plotly figures as png which is needed to use
approval testsing on figures.

* tests: add PlotlyDiffReporter for approvaltests

add custom PlotlyReporter for approvaltests to verify the image
representation of Plotly figures using diff tools which can display
images (E.g. pycharm and vscode)

* tests: verify figure using approval testing

Use ImageReporter to verify result figures by displaying an received and
approved images of the figure in a diff tool

* tests: add missing vcr decorator

* tests: remove unnecessary skip condition

* tests: fall back to reporting json if png does not work

* tests: recreate all vcr cassettes

* fix: set locale for currentness indicator

Currentness indicator does format time to string representation which
where different depending on the locale setting of the machine. This
commit will set the locale to always be en_US.

* tests(approval): scrub fitted values

for mapping saturation indicator

* tests: check result description not indicator description

* tests: improve docstring

* tests: use blocking reporter for ploty figures

* tests: fallback to PythonNativeReporter on MacOS

* tests: fallback to JSON reporting

if PlotlyFigureReport fails

* tests: fix MacOS reporters

* tests: re-enable pycharm flatpak reporter

* test(reporter): wrap report func to detect failure

of call to `open` on MacOS

* test: use verify_as_json for result figures